big update, changed the project to use Visual Studio 2019 and its toolset, also removed included dependences, the user is now responsible for linking and installing their own deps.

+++ SerenityLib/Engine/src/core/Scene.cpp
@@ -255,7 +255,7 @@ Scene::~Scene(){
     unregisterEvent(EventType::SceneChanged);
     m_i->_destruct();
 }
-void Scene::update(const float& dt){
+void Scene::update(const double& dt){
 }
 glm::vec3 Scene::getBackgroundColor(){ 
     return m_i->m_BackgroundColor;
